♻️refactor: Common 에 중복 정의된 payload 리팩토링
- Payload 중복 정의 제거
- common/src/types/index.ts 리팩토링
